implement Comparable on BytesReference and share compare code between compareTo and equals
